refactor(result): change to a new module
