+++ src/middleware.rs
@@ -3,9 +3,10 @@ use std::sync::Arc;
 use futures::Future;
 
 use crate::{
-    handler::{InternalResult, Result, Runner},
+    handler::{InternalResult, Runner},
     request::Request,
     response::Response,
+    result::Result,
 };
 
 impl From<Request<String>> for InternalResult<Request<String>> {
@@ -184,11 +185,8 @@ mod test {
     use std::sync::Arc;
 
     use crate::{
-        error::Error,
-        handler::{Result, Runner},
-        middleware::MiddlewareFactory,
-        request::Request,
-        response::Response,
+        error::Error, handler::Runner, middleware::MiddlewareFactory, request::Request,
+        response::Response, result::Result,
     };
 
     async fn pre_middleware(_req: Result<Request<String>>) -> Result<Request<String>> {
